feat: move common controllers methods to helpers and add types
