+++ src/controllers/controller.scatter.js
@@ -1,24 +1,17 @@
 import DatasetController from '../core/core.datasetController';
 import {isNullOrUndef} from '../helpers';
-import {_limitValue, isNumber} from '../helpers/helpers.math';
-import {_lookupByKey} from '../helpers/helpers.collection';
+import {isNumber} from '../helpers/helpers.math';
+import {getStartAndCountOfVisiblePoints, scaleRangesChanged} from '../helpers/helpers.extras';
+import registry from '../core/core.registry';
 
 export default class ScatterController extends DatasetController {
   initialize() {
-    super.initialize();
-
-    const {dataset} = this._cachedMeta || null;
-    const {showLine} = this.options || false;
-
-    if (showLine && dataset === null) {
-      const config = {
-        datasetElementType: 'line',
-        dataElementType: 'point',
-        showLine: true,
-        fill: false,
-      };
-      this.chart.buildOrUpdateControllers(config);
+    const {showLine} = this.getDataset();
+    if (!this.datasetElementType && showLine) {
+      this.datasetElementType = registry.getElement('line');
     }
+
+    super.initialize();
   }
 
   update(mode) {
@@ -173,54 +166,3 @@ ScatterController.overrides = {
     }
   }
 };
-
-function getStartAndCountOfVisiblePoints(meta, points, animationsDisabled) {
-  const pointCount = points.length;
-
-  let start = 0;
-  let count = pointCount;
-
-  if (meta._sorted) {
-    const {iScale, _parsed} = meta;
-    const axis = iScale.axis;
-    const {min, max, minDefined, maxDefined} = iScale.getUserBounds();
-
-    if (minDefined) {
-      start = _limitValue(Math.min(
-        _lookupByKey(_parsed, iScale.axis, min).lo,
-        animationsDisabled ? pointCount : _lookupByKey(points, axis, iScale.getPixelForValue(min)).lo),
-      0, pointCount - 1);
-    }
-    if (maxDefined) {
-      count = _limitValue(Math.max(
-        _lookupByKey(_parsed, iScale.axis, max).hi + 1,
-        animationsDisabled ? 0 : _lookupByKey(points, axis, iScale.getPixelForValue(max)).hi + 1),
-      start, pointCount) - start;
-    } else {
-      count = pointCount - start;
-    }
-  }
-
-  return {start, count};
-}
-
-function scaleRangesChanged(meta) {
-  const {xScale, yScale, _scaleRanges} = meta;
-  const newRanges = {
-    xmin: xScale.min,
-    xmax: xScale.max,
-    ymin: yScale.min,
-    ymax: yScale.max
-  };
-  if (!_scaleRanges) {
-    meta._scaleRanges = newRanges;
-    return true;
-  }
-  const changed = _scaleRanges.xmin !== xScale.min
-		|| _scaleRanges.xmax !== xScale.max
-		|| _scaleRanges.ymin !== yScale.min
-		|| _scaleRanges.ymax !== yScale.max;
-
-  Object.assign(_scaleRanges, newRanges);
-  return changed;
-}
